Copy all .md files to output folder.
